+++ internal/service/file_test.go
@@ -29,7 +29,7 @@ func setupFileTest() MockFileServices {
 	mockFolderService := new(mocks.FolderService)
 	mockLocalFileSystemService := new(mocks.LocalFileSystemService)
 
-	fileService := NewFileService(mockLogger, mockCache, mockFileRepo, mockFolderService, mockLocalFileSystemService, false, 100*1024*1024, 0)
+	fileService := NewFileService(mockLogger, mockCache, mockFileRepo, mockFolderService, mockLocalFileSystemService, false, 100*1024*1024, 0, nil)
 
 	return MockFileServices{
 		fileService:    fileService,